scripts/requirements.txt
supabase>=2.0.0
python-dotenv>=1.0.0
openpyxl>=3.1.0
pandas>=2.0.0

